create a separate deck class that contains methods specific to decks. The CardGame class will have a default constructor that generates a deck
