+++ src/test/java/io/zipcoder/casino/GameToolsTests/Deck/DeckTest.java
@@ -1,0 +1,4 @@
+package io.zipcoder.casino.GameToolsTests.Deck;
+
+public class DeckTest {
+}
